add correct price formatting for product prices showing on home screen

+++ hawkeye-crafts/Models/PlantStoolModel.swift
@@ -12,5 +12,5 @@ struct PlantStool: Identifiable {
     var id = UUID()
     var title: String
     var image: String
-    var price: Float
+    var price: Decimal
 }
